Introduces Collection Box functionality
Adds the functionality to manage collection boxes for fundraising events.
This includes creating collection boxes and associating them with events.
It also removes the `registered` field from the CollectionBox model as it
was deemed unnecessary.
A dedicated service layer is introduced for managing both
FundraisingEvent and CollectionBox entities, promoting separation of concerns
and encapsulating business logic.